+++ drift/lib/src/sentry_span_helper.dart
@@ -1,53 +1,37 @@
-import 'package:meta/meta.dart';
+// ignore_for_file: invalid_use_of_internal_member
 
+import 'package:meta/meta.dart';
 import 'package:sentry/sentry.dart';
 
-import 'sentry_query_executor.dart';
-
-/// @nodoc
 @internal
 class SentrySpanHelper {
-  /// @nodoc
-  Hub _hub = HubAdapter();
-
-  /// @nodoc
+  final Hub _hub;
   final String _origin;
+  ISentrySpan? _parentSpan;
 
-  /// @nodoc
-  SentrySpanHelper(this._origin);
+  SentrySpanHelper(this._origin, {Hub? hub}) : _hub = hub ?? HubAdapter();
 
-  /// @nodoc
-  void setHub(Hub hub) {
-    _hub = hub;
-  }
-
-  /// @nodoc
-  @internal
   Future<T> asyncWrapInSpan<T>(
     String description,
     Future<T> Function() execute, {
     String? dbName,
-    bool useTransactionSpan = false,
+    String? operation,
   }) async {
-    ISentrySpan? currentSpan = _hub.getSpan();
-    if (useTransactionSpan) {
-      currentSpan = transactionSpan;
-    }
+    final currentSpan = _parentSpan ?? _hub.getSpan();
     final span = currentSpan?.startChild(
-      SentryQueryExecutor.dbOp,
+      operation ?? SentrySpanOperations.dbSqlQuery,
       description: description,
     );
 
-    // ignore: invalid_use_of_internal_member
     span?.origin = _origin;
 
     span?.setData(
-      SentryQueryExecutor.dbSystemKey,
-      SentryQueryExecutor.dbSystem,
+      SentrySpanData.dbSystemKey,
+      SentrySpanData.dbSystemSqlite,
     );
 
     if (dbName != null) {
-      span?.setData(SentryQueryExecutor.dbNameKey, dbName);
+      span?.setData(SentrySpanData.dbNameKey, dbName);
     }
 
     try {
@@ -65,91 +49,71 @@ class SentrySpanHelper {
     }
   }
 
-  /// This span is used for the database transaction.
-  @internal
-  ISentrySpan? transactionSpan;
-
-  /// @nodoc
-  @internal
   T beginTransaction<T>(
-    String description,
     T Function() execute, {
     String? dbName,
   }) {
-    final currentSpan = _hub.getSpan();
-    final span = currentSpan?.startChild(
-      SentryQueryExecutor.dbOp,
-      description: description,
+    final scopeSpan = _hub.getSpan();
+    _parentSpan = scopeSpan?.startChild(
+      SentrySpanOperations.dbSqlTransaction,
+      description: SentrySpanDescriptions.dbTransaction,
     );
 
-    // ignore: invalid_use_of_internal_member
-    span?.origin = _origin;
+    _parentSpan?.origin = _origin;
 
-    span?.setData(
-      SentryQueryExecutor.dbSystemKey,
-      SentryQueryExecutor.dbSystem,
+    _parentSpan?.setData(
+      SentrySpanData.dbSystemKey,
+      SentrySpanData.dbSystemSqlite,
     );
 
     if (dbName != null) {
-      span?.setData(SentryQueryExecutor.dbNameKey, dbName);
+      _parentSpan?.setData(SentrySpanData.dbNameKey, dbName);
     }
 
     try {
       final result = execute();
-      span?.status = SpanStatus.unknown();
+      _parentSpan?.status = SpanStatus.unknown();
 
       return result;
     } catch (exception) {
-      span?.throwable = exception;
-      span?.status = SpanStatus.internalError();
+      _parentSpan?.throwable = exception;
+      _parentSpan?.status = SpanStatus.internalError();
 
       rethrow;
-    } finally {
-      transactionSpan = span;
     }
   }
 
-  /// @nodoc
-  @internal
-  Future<T> finishTransaction<T>(
-    Future<T> Function() execute, {
-    String? dbName,
-  }) async {
+  Future<T> finishTransaction<T>(Future<T> Function() execute) async {
     try {
       final result = await execute();
-      transactionSpan?.status = SpanStatus.ok();
+      _parentSpan?.status = SpanStatus.ok();
 
       return result;
     } catch (exception) {
-      transactionSpan?.throwable = exception;
-      transactionSpan?.status = SpanStatus.internalError();
+      _parentSpan?.throwable = exception;
+      _parentSpan?.status = SpanStatus.internalError();
 
       rethrow;
     } finally {
-      await transactionSpan?.finish();
-      transactionSpan = null;
+      await _parentSpan?.finish();
+      _parentSpan = null;
     }
   }
 
-  /// @nodoc
-  @internal
-  Future<T> abortTransaction<T>(
-    Future<T> Function() execute, {
-    String? dbName,
-  }) async {
+  Future<T> abortTransaction<T>(Future<T> Function() execute) async {
     try {
       final result = await execute();
-      transactionSpan?.status = SpanStatus.aborted();
+      _parentSpan?.status = SpanStatus.aborted();
 
       return result;
     } catch (exception) {
-      transactionSpan?.throwable = exception;
-      transactionSpan?.status = SpanStatus.internalError();
+      _parentSpan?.throwable = exception;
+      _parentSpan?.status = SpanStatus.internalError();
 
       rethrow;
     } finally {
-      await transactionSpan?.finish();
-      transactionSpan = null;
+      await _parentSpan?.finish();
+      _parentSpan = null;
     }
   }
 }