+++ drift/test/sentry_drift_test.dart
@@ -8,8 +8,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/sentry_tracer.dart';
-import 'package:sentry_drift/src/sentry_query_executor.dart';
-import 'package:sentry_drift/src/sentry_transaction_executor.dart';
+import 'package:sentry_drift/sentry_drift.dart';
+import 'package:sentry_drift/src/constants.dart' as drift_constants;
 import 'package:sentry_drift/src/version.dart';
 import 'package:sqlite3/open.dart';
 
@@ -27,28 +27,27 @@ void main() {
       'UPDATE "todo_items" SET "title" = ?, "body" = ? WHERE "title" = ?;';
   final expectedSelectStatement = 'SELECT * FROM todo_items';
   final expectedDeleteStatement = 'DELETE FROM "todo_items";';
-  final expectedCloseStatement = 'Close DB: ${Fixture.dbName}';
-  final expectedOpenStatement = 'Open DB: ${Fixture.dbName}';
-  final expectedTransactionStatement = 'transaction';
-  final withinTransactionDescription = 'Within transaction: ';
 
   void verifySpan(
     String description,
     SentrySpan? span, {
-    String origin = SentryTraceOrigins.autoDbDriftQueryExecutor,
+    String? operation,
     SpanStatus? status,
   }) {
     status ??= SpanStatus.ok();
-    expect(span?.context.operation, SentryQueryExecutor.dbOp);
+    expect(
+      span?.context.operation,
+      operation ?? SentrySpanOperations.dbSqlQuery,
+    );
     expect(span?.context.description, description);
     expect(span?.status, status);
-    expect(span?.origin, origin);
+    expect(span?.origin, SentryTraceOrigins.autoDbDriftQueryInterceptor);
     expect(
-      span?.data[SentryQueryExecutor.dbSystemKey],
-      SentryQueryExecutor.dbSystem,
+      span?.data[SentrySpanData.dbSystemKey],
+      SentrySpanData.dbSystemSqlite,
     );
     expect(
-      span?.data[SentryQueryExecutor.dbNameKey],
+      span?.data[SentrySpanData.dbNameKey],
       Fixture.dbName,
     );
   }
@@ -57,18 +56,22 @@ void main() {
     String description,
     Exception exception,
     SentrySpan? span, {
-    String origin = SentryTraceOrigins.autoDbDriftQueryExecutor,
+    String? operation,
+    SpanStatus? status,
   }) {
-    expect(span?.context.operation, SentryQueryExecutor.dbOp);
-    expect(span?.context.description, description);
-    expect(span?.status, SpanStatus.internalError());
-    expect(span?.origin, origin);
     expect(
-      span?.data[SentryQueryExecutor.dbSystemKey],
-      SentryQueryExecutor.dbSystem,
+      span?.context.operation,
+      operation ?? SentrySpanOperations.dbSqlQuery,
+    );
+    expect(span?.context.description, description);
+    expect(span?.status, status ?? SpanStatus.internalError());
+    expect(span?.origin, SentryTraceOrigins.autoDbDriftQueryInterceptor);
+    expect(
+      span?.data[SentrySpanData.dbSystemKey],
+      SentrySpanData.dbSystemSqlite,
     );
     expect(
-      span?.data[SentryQueryExecutor.dbNameKey],
+      span?.data[SentrySpanData.dbNameKey],
       Fixture.dbName,
     );
 
@@ -91,6 +94,21 @@ void main() {
             ),
           );
     }
+  }
+
+  Future<void> insertIntoBatch(AppDatabase sut) {
+    return sut.batch((batch) {
+      batch.insertAll(sut.todoItems, [
+        TodoItemsCompanion.insert(
+          title: 'todo: finish drift setup #1',
+          content: 'We can now write queries and define our own tables.',
+        ),
+        TodoItemsCompanion.insert(
+          title: 'todo: finish drift setup #2',
+          content: 'We can now write queries and define our own tables.',
+        ),
+      ]);
+    });
   }
 
   Future<void> updateRow(AppDatabase sut, {bool withError = false}) {
@@ -140,7 +158,9 @@ void main() {
 
       final openSpansCount = fixture.tracer.children
           .where(
-            (element) => element.context.description == expectedOpenStatement,
+            (element) =>
+                element.context.description ==
+                SentrySpanDescriptions.dbOpen(dbName: Fixture.dbName),
           )
           .length;
 
@@ -203,23 +223,21 @@ void main() {
 
       final insertSpanCount = fixture.tracer.children
           .where(
-            (element) =>
-                element.context.description ==
-                '$withinTransactionDescription$expectedInsertStatement',
+            (element) => element.context.description == expectedInsertStatement,
           )
           .length;
       expect(insertSpanCount, 2);
 
       verifySpan(
-        '$withinTransactionDescription$expectedInsertStatement',
+        expectedInsertStatement,
         fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
       );
 
       verifySpan(
-        expectedTransactionStatement,
-        fixture.getCreatedSpanByDescription(expectedTransactionStatement),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        SentrySpanDescriptions.dbTransaction,
+        fixture
+            .getCreatedSpanByDescription(SentrySpanDescriptions.dbTransaction),
+        operation: SentrySpanOperations.dbSqlTransaction,
       );
     });
 
@@ -233,23 +251,21 @@ void main() {
 
       final updateSpanCount = fixture.tracer.children
           .where(
-            (element) =>
-                element.context.description ==
-                '$withinTransactionDescription$expectedUpdateStatement',
+            (element) => element.context.description == expectedUpdateStatement,
           )
           .length;
       expect(updateSpanCount, 1);
 
       verifySpan(
-        '$withinTransactionDescription$expectedUpdateStatement',
+        expectedUpdateStatement,
         fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
       );
 
       verifySpan(
-        expectedTransactionStatement,
-        fixture.getCreatedSpanByDescription(expectedTransactionStatement),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        SentrySpanDescriptions.dbTransaction,
+        fixture
+            .getCreatedSpanByDescription(SentrySpanDescriptions.dbTransaction),
+        operation: SentrySpanOperations.dbSqlTransaction,
       );
     });
 
@@ -263,23 +279,21 @@ void main() {
 
       final deleteSpanCount = fixture.tracer.children
           .where(
-            (element) =>
-                element.context.description ==
-                '$withinTransactionDescription$expectedDeleteStatement',
+            (element) => element.context.description == expectedDeleteStatement,
           )
           .length;
       expect(deleteSpanCount, 1);
 
       verifySpan(
-        '$withinTransactionDescription$expectedDeleteStatement',
+        expectedDeleteStatement,
         fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
       );
 
       verifySpan(
-        expectedTransactionStatement,
-        fixture.getCreatedSpanByDescription(expectedTransactionStatement),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        SentrySpanDescriptions.dbTransaction,
+        fixture
+            .getCreatedSpanByDescription(SentrySpanDescriptions.dbTransaction),
+        operation: SentrySpanOperations.dbSqlTransaction,
       );
     });
 
@@ -293,23 +307,21 @@ void main() {
 
       final customSpanCount = fixture.tracer.children
           .where(
-            (element) =>
-                element.context.description ==
-                '$withinTransactionDescription$expectedSelectStatement',
+            (element) => element.context.description == expectedSelectStatement,
           )
           .length;
       expect(customSpanCount, 1);
 
       verifySpan(
-        '$withinTransactionDescription$expectedSelectStatement',
+        expectedSelectStatement,
         fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
       );
 
       verifySpan(
-        expectedTransactionStatement,
-        fixture.getCreatedSpanByDescription(expectedTransactionStatement),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        SentrySpanDescriptions.dbTransaction,
+        fixture
+            .getCreatedSpanByDescription(SentrySpanDescriptions.dbTransaction),
+        operation: SentrySpanOperations.dbSqlTransaction,
       );
     });
 
@@ -331,25 +343,22 @@ void main() {
       final abortedSpan = spans.first;
 
       verifySpan(
-        expectedTransactionStatement,
+        SentrySpanDescriptions.dbTransaction,
         abortedSpan,
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
         status: SpanStatus.aborted(),
+        operation: SentrySpanOperations.dbSqlTransaction,
       );
     });
 
     test('batch adds span', () async {
       final sut = fixture.sut;
 
-      await sut.batch((batch) async {
-        await insertRow(sut);
-        await insertRow(sut);
-      });
+      await insertIntoBatch(sut);
 
       verifySpan(
-        'batch',
+        SentrySpanDescriptions.dbBatch(statements: [expectedInsertStatement]),
         fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        operation: SentrySpanOperations.dbSqlBatch,
       );
     });
 
@@ -359,8 +368,9 @@ void main() {
       await sut.close();
 
       verifySpan(
-        'Close DB: ${Fixture.dbName}',
+        SentrySpanDescriptions.dbClose(dbName: Fixture.dbName),
         fixture.getCreatedSpan(),
+        operation: SentrySpanOperations.dbClose,
       );
     });
 
@@ -372,8 +382,11 @@ void main() {
       await sut.select(sut.todoItems).get();
 
       verifySpan(
-        expectedOpenStatement,
-        fixture.getCreatedSpanByDescription(expectedOpenStatement),
+        SentrySpanDescriptions.dbOpen(dbName: Fixture.dbName),
+        fixture.getCreatedSpanByDescription(
+          SentrySpanDescriptions.dbOpen(dbName: Fixture.dbName),
+        ),
+        operation: SentrySpanOperations.dbOpen,
       );
     });
   });
@@ -424,15 +437,19 @@ void main() {
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
       when(fixture.mockLazyDatabase.ensureOpen(any))
           .thenAnswer((_) => Future.value(true));
-
-      await fixture.setUp(injectMock: true);
+      when(fixture.mockLazyDatabase.dialect).thenReturn(SqlDialect.sqlite);
     });
 
     tearDown(() async {
-      await fixture.tearDown();
+      // catch errors because we purposefully throw a close in one of the tests
+      try {
+        await fixture.tearDown();
+      } catch (_) {}
     });
 
     test('throwing runInsert throws error span', () async {
+      await fixture.setUp(injectMock: true);
+
       when(fixture.mockLazyDatabase.runInsert(any, any))
           .thenThrow(fixture.exception);
 
@@ -450,8 +467,12 @@ void main() {
     });
 
     test('throwing runUpdate throws error span', () async {
+      await fixture.setUp(injectMock: true);
+
       when(fixture.mockLazyDatabase.runUpdate(any, any))
           .thenThrow(fixture.exception);
+      when(fixture.mockLazyDatabase.ensureOpen(any))
+          .thenAnswer((_) => Future.value(true));
 
       try {
         await updateRow(fixture.sut);
@@ -467,6 +488,8 @@ void main() {
     });
 
     test('throwing runCustom throws error span', () async {
+      await fixture.setUp(injectMock: true);
+
       when(fixture.mockLazyDatabase.runCustom(any, any))
           .thenThrow(fixture.exception);
 
@@ -487,20 +510,12 @@ void main() {
       final mockTransactionExecutor = MockTransactionExecutor();
       when(mockTransactionExecutor.beginTransaction())
           .thenThrow(fixture.exception);
+      when(mockTransactionExecutor.ensureOpen(any))
+          .thenAnswer((_) => Future.value(true));
+
+      await fixture.setUp(customExecutor: mockTransactionExecutor);
 
       try {
-        // We need to move it inside the try/catch becaue SentryTransactionExecutor
-        // starts beginTransaction() directly after init
-        final SentryTransactionExecutor transactionExecutor =
-            SentryTransactionExecutor(
-          mockTransactionExecutor,
-          fixture.hub,
-          dbName: Fixture.dbName,
-        );
-
-        when(fixture.mockLazyDatabase.beginTransaction())
-            .thenReturn(transactionExecutor);
-
         await fixture.sut.transaction(() async {
           await insertRow(fixture.sut);
         });
@@ -509,46 +524,48 @@ void main() {
       }
 
       verifyErrorSpan(
-        expectedTransactionStatement,
+        SentrySpanDescriptions.dbTransaction,
         fixture.exception,
         fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        operation: SentrySpanOperations.dbSqlTransaction,
       );
     });
 
-    test('throwing batch throws error span', () async {
+    test('throwing batch throws error span in transaction', () async {
+      await fixture.setUp(injectMock: true);
+
       final mockTransactionExecutor = MockTransactionExecutor();
-      when(mockTransactionExecutor.beginTransaction())
+      when(mockTransactionExecutor.ensureOpen(any))
+          .thenAnswer((_) => Future.value(true));
+      when(fixture.mockLazyDatabase.beginTransaction())
+          .thenReturn(mockTransactionExecutor);
+      when(mockTransactionExecutor.runBatched(any))
           .thenThrow(fixture.exception);
 
-      // We need to move it inside the try/catch becaue SentryTransactionExecutor
-      // starts beginTransaction() directly after init
-      final SentryTransactionExecutor transactionExecutor =
-          SentryTransactionExecutor(
-        mockTransactionExecutor,
-        fixture.hub,
-        dbName: Fixture.dbName,
+      await expectLater(
+        () async => await insertIntoBatch(fixture.sut),
+        throwsException,
       );
 
-      when(fixture.mockLazyDatabase.beginTransaction())
-          .thenReturn(transactionExecutor);
-
-      when(fixture.mockLazyDatabase.runInsert(any, any))
-          .thenAnswer((realInvocation) => Future.value(1));
-
-      try {
-        await fixture.sut.batch((batch) async {
-          await insertRow(fixture.sut);
-        });
-      } catch (exception) {
-        expect(exception, fixture.exception);
-      }
-
+      // errored batch
       verifyErrorSpan(
-        expectedTransactionStatement,
+        SentrySpanDescriptions.dbBatch(statements: [expectedInsertStatement]),
         fixture.exception,
-        fixture.getCreatedSpan(),
-        origin: SentryTraceOrigins.autoDbDriftTransactionExecutor,
+        fixture.getCreatedSpanByDescription(
+          SentrySpanDescriptions.dbBatch(
+            statements: [expectedInsertStatement],
+          ),
+        ),
+        operation: SentrySpanOperations.dbSqlBatch,
+      );
+
+      // // aborted transaction
+      verifySpan(
+        SentrySpanDescriptions.dbTransaction,
+        fixture
+            .getCreatedSpanByDescription(SentrySpanDescriptions.dbTransaction),
+        operation: SentrySpanOperations.dbSqlTransaction,
+        status: SpanStatus.aborted(),
       );
     });
 
@@ -556,6 +573,8 @@ void main() {
       when(fixture.mockLazyDatabase.close()).thenThrow(fixture.exception);
       when(fixture.mockLazyDatabase.runInsert(any, any))
           .thenAnswer((_) => Future.value(1));
+
+      await fixture.setUp(injectMock: true);
 
       try {
         await insertRow(fixture.sut);
@@ -565,15 +584,18 @@ void main() {
       }
 
       verifyErrorSpan(
-        expectedCloseStatement,
+        SentrySpanDescriptions.dbClose(dbName: Fixture.dbName),
         fixture.exception,
-        fixture.getCreatedSpan(),
+        fixture.getCreatedSpanByDescription(
+          SentrySpanDescriptions.dbClose(dbName: Fixture.dbName),
+        ),
+        operation: SentrySpanOperations.dbClose,
       );
-
-      when(fixture.mockLazyDatabase.close()).thenAnswer((_) => Future.value());
     });
 
     test('throwing ensureOpen throws error span', () async {
+      await fixture.setUp(injectMock: true);
+
       when(fixture.mockLazyDatabase.ensureOpen(any))
           .thenThrow(fixture.exception);
 
@@ -584,13 +606,18 @@ void main() {
       }
 
       verifyErrorSpan(
-        expectedOpenStatement,
+        SentrySpanDescriptions.dbOpen(dbName: Fixture.dbName),
         fixture.exception,
-        fixture.getCreatedSpanByDescription(expectedOpenStatement),
+        fixture.getCreatedSpanByDescription(
+          SentrySpanDescriptions.dbOpen(dbName: Fixture.dbName),
+        ),
+        operation: SentrySpanOperations.dbOpen,
       );
     });
 
     test('throwing runDelete throws error span', () async {
+      await fixture.setUp(injectMock: true);
+
       when(fixture.mockLazyDatabase.runDelete(any, any))
           .thenThrow(fixture.exception);
 
@@ -626,7 +653,8 @@ void main() {
 
     test('adds integration', () {
       expect(
-        fixture.options.sdk.integrations.contains('SentryDriftTracing'),
+        fixture.options.sdk.integrations
+            .contains(drift_constants.integrationName),
         true,
       );
     });
@@ -653,8 +681,14 @@ class Fixture {
   late AppDatabase sut;
   final mockLazyDatabase = MockLazyDatabase();
 
-  Future<void> setUp({bool injectMock = false}) async {
-    sut = AppDatabase(openConnection(injectMock: injectMock));
+  Future<void> setUp({
+    bool injectMock = false,
+    QueryExecutor? customExecutor,
+  }) async {
+    sut = AppDatabase(
+      openConnection(injectMock: injectMock, customExecutor: customExecutor),
+    );
+    driftRuntimeOptions.dontWarnAboutMultipleDatabases = true;
   }
 
   Future<void> tearDown() async {
@@ -670,19 +704,21 @@ class Fixture {
         .firstWhere((element) => element.context.description == description);
   }
 
-  SentryQueryExecutor openConnection({bool injectMock = false}) {
-    if (injectMock) {
-      final executor =
-          SentryQueryExecutor(() => mockLazyDatabase, databaseName: dbName);
-      executor.setHub(hub);
-      return executor;
+  QueryExecutor openConnection({
+    bool injectMock = false,
+    QueryExecutor? customExecutor,
+  }) {
+    if (customExecutor != null) {
+      return customExecutor.interceptWith(
+        SentryQueryInterceptor(databaseName: dbName, hub: hub),
+      );
+    } else if (injectMock) {
+      return mockLazyDatabase.interceptWith(
+        SentryQueryInterceptor(databaseName: dbName, hub: hub),
+      );
     } else {
-      return SentryQueryExecutor(
-        () {
-          return NativeDatabase.memory();
-        },
-        hub: hub,
-        databaseName: dbName,
+      return NativeDatabase.memory().interceptWith(
+        SentryQueryInterceptor(databaseName: dbName, hub: hub),
       );
     }
   }